also completed the indexOf(obj) method
looping through nodes logic is same as most of the methods, using similar track variable of int type (like that in remove method) to keep track of the index. and remove it if the current string matches the obj parameter passed inside it. returning -1 as that is how i saw the api displays if string at that index is not found.