Figured out the way to make methods generic
FInally found the way to make the methods generic, took so long. FOund out that i need to use override and keep the parameter same to be able to make it work, but inside change the paramter typ to T(so that way it can store any type of data). Copied the whole add method from DSEList class, and did these changes.